rename tests & simplify transactions test
