+++ contracts/scripts/e2e/bridge/standard_bridge_erc20.ts
@@ -7,7 +7,7 @@ import {
   delay,
   getLastBlockNumber,
   deployTokenPair,
-} from "./utils";
+} from "../utils";
 
 async function main() {
   const {
