+++ contracts/scripts/e2e/bridge/standard_bridge_withdraw_eth.ts
@@ -5,7 +5,7 @@ import {
   getWithdrawalProof,
   delay,
   getLastBlockNumber,
-} from "./utils";
+} from "../utils";
 
 async function main() {
   const {
